+++ wallet/core/src/wasm/wallet/account.rs
@@ -2,11 +2,12 @@ use crate::account as native;
 use crate::imports::*;
 use crate::tx::PaymentOutputs;
 use crate::wasm::utxo::UtxoContext;
+use kaspa_consensus_core::network::INetworkType;
 use kaspa_wallet_keys::keypair::Keypair;
 use workflow_core::abortable::Abortable;
 use workflow_wasm::abi::ref_from_abi;
 
-/// @category Wallet SDK
+/// @category Wallet API
 #[wasm_bindgen(inspectable)]
 #[derive(Clone)]
 pub struct Account {
@@ -46,9 +47,9 @@ impl Account {
     }
 
     #[wasm_bindgen(js_name = balanceStrings)]
-    pub fn balance_strings(&self, network_type: JsValue) -> Result<JsValue> {
+    pub fn balance_strings(&self, network_type: INetworkType) -> Result<JsValue> {
         match self.inner.balance() {
-            Some(balance) => Ok(crate::wasm::Balance::from(balance).as_strings(network_type)?.into()),
+            Some(balance) => Ok(crate::wasm::Balance::from(balance).to_balance_strings(network_type)?.into()),
             None => Ok(JsValue::UNDEFINED),
         }
     }