+++ wasm/examples/nodejs/javascript/transactions/utxo-context-listener.js
@@ -35,12 +35,13 @@ const { encoding, networkId, address : destinationAddress } = require("../utils"
     });
 
     // 2) Create UtxoProcessor, passing RPC to it
-    let processor = await new UtxoProcessor({ rpc, networkId });
+    let processor = new UtxoProcessor({ rpc, networkId });
+    await processor.start();
 
     // 3) Create one of more UtxoContext, passing UtxoProcessor to it
     // you can create UtxoContext objects as needed to monitor different
     // address sets.
-    let context = await new UtxoContext({ processor });
+    let context = new UtxoContext({ processor });
 
     // 4) Register a listener with the UtxoProcessor::events
     processor.addEventListener("*", (event) => {